Fix initial position of the cat, and going from Start to Game Over screen
Now, left side of the cat always stays on the screen (added a bit of padding to the left). And, moved Y coord of the cat a bit above the center to get out of the bottom current.

Also fixed going straight from Start to Game Over screen if enough time passes (changed Cat's HungerTimer's AutoStart to be false).